rework how connections/sockets are handles
connection method is now always called in its own thread, for the sake
of consistency.

when reconnecting, store the timer in a class property so it can be
cancelled if necessary.

throw an OSError when empty binary data is received from the socket.
this happens the server closes our socket, for example because of too
many (re)connects in a short timespan.

increase the reconnect delay when throttled to 60 seconds.

don't try to send a quit message if self.irc_socket is None.

closes #9